+++ src/main/java/persistence/sql/ddl/domain/Column.java
@@ -1,5 +1,4 @@
-package persistence.domain;
-
+package persistence.sql.ddl.domain;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -8,5 +7,7 @@ import java.lang.annotation.Target;
 
 @Target({ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface Id {
+public @interface Column {
+    String name() default "";
+    boolean nullable() default true;
 }
