+++ src/main/java/persistence/sql/ddl/domain/GeneratedValue.java
@@ -1,12 +1,13 @@
-package persistence.domain;
+package persistence.sql.ddl.domain;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
+@Target({ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.TYPE)
-public @interface Table {
-    String name();
+public @interface GeneratedValue {
+    GenerationType strategy() default GenerationType.AUTO;
+
 }
