+++ src/main/java/persistence/sql/ddl/domain/Person1.java
@@ -1,4 +1,4 @@
-package persistence.domain;
+package persistence.sql.ddl.domain;
 
 
 @Entity
