+++ src/main/java/persistence/sql/ddl/domain/Person2.java
@@ -1,18 +1,16 @@
-package persistence.domain;
+package persistence.sql.ddl.domain;
 
-
-@Table(name = "users")
 @Entity
-public class Person3 {
+public class Person2 {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
+
     @Column(name = "old")
     private Integer age;
     @Column(nullable = false)
     private String email;
     @Column(name = "nick_name")
     private String name;
-    @Transient
-    private Integer index;
+
 }
