+++ src/main/java/persistence/sql/ddl/domain/Person3.java
@@ -1,16 +1,18 @@
-package persistence.domain;
+package persistence.sql.ddl.domain;
 
+
+@Table(name = "users")
 @Entity
-public class Person2 {
+public class Person3 {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-
     @Column(name = "old")
     private Integer age;
     @Column(nullable = false)
     private String email;
     @Column(name = "nick_name")
     private String name;
-
+    @Transient
+    private Integer index;
 }
