+++ src/main/java/persistence/sql/ddl/domain/Table.java
@@ -1,11 +1,12 @@
-package persistence.domain;
+package persistence.sql.ddl.domain;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-@Target({ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface Transient {
+@Target(ElementType.TYPE)
+public @interface Table {
+    String name();
 }
