+++ src/main/java/persistence/sql/ddl/domain/Transient.java
@@ -1,4 +1,4 @@
-package persistence.domain;
+package persistence.sql.ddl.domain;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,7 +7,5 @@ import java.lang.annotation.Target;
 
 @Target({ElementType.FIELD})
 @Retention(RetentionPolicy.RUNTIME)
-public @interface Column {
-    String name() default "";
-    boolean nullable() default true;
+public @interface Transient {
 }
